pyproject.toml
[project]
name = "fairy-tales-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.0",
    "pytest>=8.3.5",
    "sqlalchemy>=2.0.40",
    "alembic>=1.15.2",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "google-cloud-storage>=3.1.1",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "cachetools>=5.5.2",
    "pandas>=2.3.0",
    "pre-commit>=4.2.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "email-validator>=2.1.0",
    "openai>=1.58.1",
]
